Expose a bash shell in a remote-like environment (#8)
* Expose a bash shell in a remote-like environment

We need a way to access `manage.py` on prod/staging, but Fargate doesn't
allow you to enter a shell of a running task. Making things more
complicated, there may be instances where we want to run code that has
not yet been deployed to staging or prod (such as a local version that
has migrations.)

Adding even more complexity, it's possible multiple tags can be running
in prod or staging at the same time, as you could be accessing the
cluster in the middle of a red/black deployment. Or your session may
block a deployment.

This code allows AWS users on the VPN with the proper permissions to
connect to RDS/Redis/etc using the same environment variables as
prod/staging.

We control access to our systems 2 ways: (1) you need to be on the VPN
to access these resources, and (2) your local IAM role needs to have
permission to access and decrypt the relevant configuration parameters
for staging and prod.

As this will be quite slow (each database and redis request has to go
through the public internet through the VPN), so this is a good script
to run on a EC2 bastion located on the same private subnet with similar
security groups and IAM roles as the staging/prod fargate tasks.

* Allow running version of app locally with remote settings.