Add ability to invite new users to turnout
You can now go into the django admin and create an "invite" for an email
address, which will generate a unique URL that the end-user can visit to
create their profile and set their password.

This link expires after 7 days, but the expiration time can be changed
or customized inside the django admin.