Multi Factor Authentication using Google Auth/Authy (#29)
* Multi Factor Authentication using Google Auth/Authy

This implements time-based 2 factor authentication using apps such as
Google Authenticator and Authy. It rips a LOT of stuff out of the
django_otp, but goes its own way because

(1) django_otp allows users to have multiple devices of different types
and we don't need to support that workflow yet
(2) django_otp uses auto-incrementing primary keys for Device models,
and we're trying to use UUIDs everywhere possible.

The advantage of building on top of django_otp is that we can support
features such as email and SMS one-time-codes in the future with a bit
of refactoring. Also it handles a lot of the verify-the-provided-code
bits.

I'm using a custom fork of django_otp for now, as by default django_otp
doesn't support non-auto-incrementing integers as primary keys. I
submitted my changes on my fork to the main django_otp project and once
that is merged into the main package we can likely move to that.

* Handle purest failure by removing django_otp plugin imports

There was a problem where django's app finder was finding the standard
django_otp totp plugin. Since that plugin was not in INSTALLED_APPS
django was trying to create a table from that module when initially
setting up the test database.

Removing all references to the plugin from our code appears to have
fixed the issue.