Force user's browser to download files (#64)
If a user's browser opens a PDF or image in the browser, then they
refresh or share the URL in their browser's URL bar, it will result in
an ugly S3 "this link has expired" error.

In addition, in our effort to dedupe files on S3 we choose pretty ugly
custom-per-file URLs. To keep things clean on a user's computer, we
should have users download the file with the original filename.