Allow blank information fields in admin, template cleanup (#76)
The `text` field allows blank entries, but our validator will throw
exceptions if a blank value is provided. This is particularly annoying
for URLs where we want to clear out the URL.

Also, there was a request to make the slug clickable on the information
field type page, which is how we have it on the state information page.