Add `Action` OneToOne fields, initial event tracking (#83)
* Add event tracking for Registrations and Lookups

For analytics purposes we'd like to be able to track events such as
official website visits. In order to make this reproducible we can use a
new OneToOneField which links to an `Action` object that has events
that foreign key to it. It's pretty easy to add this object to both
Registrations and Lookups, then return the action id.

After the action id is acquired, you can POST to `/v1/event/track/` the
action ID as `action` and the event type as `event_type`

Currently the only event type allowed is `OToolVisit` (or official state
tool visit) although this can be expanded on by updating the enums.

* Allow field values up to 100 characters long

We likely never want an event name to be that long

* Move generation of the Action to a serializer