Expose fields needed for disclaimers in partner embeds (#111)
* Expose fields needed for disclaimers in partner embeds

In order to support partners we may need to show their name as well as
provide links to their terms of service and privacy policy. In addition
some will need to have checkboxes that opt the user into their SMS
program. Some of those checkboxes will be checked by default while
others will not.

We can expose all of this in embeds

* Remove unnecessary mixin