official: add /v1/official/address/ endpoint to find correct region (#142)
* Makefile: make psql

* docker-compose: switch to postgis, install necessary libs

* /v1/official/address: add API to get closest LEO

Most of the time this returns a single result (exact match on county,
city, etc.).

In a few states we have some special handling for weird naming or
boundaries.  For WI, we query an arcgis API to get the exact municipality
name (village, city, town) to match against.

If necessary, we fall back to a search by lat/lng and return the 5 closest
offices.  This is usually only necessary in very small towns or
unincorporated areas in states like WI, MI, or much of New England.

* official: add location to Address

* /v1/official/address: handle AL Jefferson County divisions

Jefferson county has a weird division between Bessemer and Birmingham
county divisions.

* /v1/official/address: use postgris to get closest offices

also use select_related to get region in one query

* chore: add a couple geocode tests

* bugfix: common/geocode: fix case on WI API result

* bugfix: official/address: fix a few WI corner cases

* summarize in a doc

* common/geocode: add timeout and retry

* geocode: log errors

* refactor to use either TS or geocod.io

* improve logging and log exceptions with sentry

* cleanup a bit

* update readme