Improve accuracy of /v1/official/address (#193)
* improve accuracy of /v1/official/address/ lookup

The geocodio method tends to perform much better than targetsmart, except
for just a few states: MI and VT, where the regions are super weird.  The
TS API is simpler to use (a single lookup by address provides the
info the need), but it always fails on one of the addresses in the test
set (175 grove st, cambridge 02138) which is on the border of Cambridge
and Belmont; geocodio gets it right.  Mapping the address to lat/lng first
and passing that to TS also provides the wrong answer.

So, always try geocodio first and falling back to TS *except* for the
prefer_ts states (MI VT).

* handle county name ambiguity "LaSalle" vs "La Salle" (and LaPorte, etc)

* handle 'St. ' in mcd

* only try the first few geocod.io addresses; ignore additional addrs if accuracy drops

* only fall back to a pure location-based match if all address matches fail

* Honolulu is "Honolulu, City and County"

* misc test

* bugfix: common/geocode: fix exception handling

* fix south dakota county

* fix nyc

* handle ' in county name (Prince George's County, MD)

* DeSoto -> De Soto

* occasionally geocod.io returns no county

605+redstone+dr&city=Bristol&state=TN&zipcode=37620

* another hard MA one

* Le Flore -> LeFlore

* only check the first geocodio address

this seems to be sufficient....

* behave if geocod.io returns no county

* fix Richmond ("Richmond city" -> "City of Richmond")

* mangle addresses

* lint

* query state GIS for MA too

MA appears to be the last state where we fail to reliably map to the
correct municipality based on the geocoding results alone.  Fortunately,
everyone uses the same GIS software, and MA has their official data
set available.

* handle typo for {"name":"Ctiy of Gardner","external_id":431107}

* common/geocode: allow MA and WI precint/ward/muni data to be imported locally

Importing these locally avoids the external API call and saves ~300ms.

* put models in official/models.py; drop public API query support

* download data before dropping table and importing

Signed-off-by: Sage Weil <sage@newdream.net>

* empty migrations for new models