Have the datadog container pass along the `env` tag (#206)
In order for us to create filterable dashboards in datadog we need to
pass along the `env` tag everywhere a metric is passed along. In this
case we were not passing that env tag along from the datadog container
that is responsible for collecting metrics for the fargate task.

Spinnaker automatically adds a docker label called `spinnaker.stack`
which includes the environment name. We can tell the datadog container
to look at that docker label and pass it along as a tag called `env`.