Add additional APM tracing (#213)
* Add additional APM tracing

In order to get more visibility in the datadog flame graph we need to
wrap key methods we care about in `@tracer.wrap()`. I included anything
that may involve processing time, such as hitting external services or
rendering templates. Monitoring email compile time is going to be
especially useful if we ever decide to get more complex with our
transactional emails or throw more fields in.

The other advantage here is we'll be able to link individual queries
in the flame graph to components of the request, instead of the entire
view.

* Move tracing of targetsmart and alloy to separate services

* Remove redundant calls to statsd when the tracer wraps

We get timing by default with tracer.wrap() so there is no need to also
pass the timing information in via statsd.

* Get rid of breaking test

* lint