Only validate each field one time for every request (#218)
With the previous setup each request to the server would validate each
field more times than it did previously. So the first request would
validate the email is in the correct format one time, the second would
check the email twice (or more), etc. This was causing significant
slowness in requests, with every request taking twice as long as the
previous one.

This can be fixed by bypassing some unnecessary default django rest
framework code that generates what fields should be included with each
request.