Add hidden submit button to allow better logging in (#256)
Because the "Back" button was the first button the user's browser saw
after the token entering form, anyone who pressed enter would be
immediately sent back to the 1st authentication page. If someone was
a next-button-clicker they'd be fine, but anyone who was down with
keyboards was having a weird relationship with our login flow.

Also added some styles to the flow, because things should look nice.