Convert SmallUUID to UUID (#314)
* Move from SmallUUID to UUID

It's extordinarially difficult to run queries using the raw database
then login to the admin and find those records. Alternatively, it's
difficult to find a record in the admin then find associated records in
the database.

By switching from SmallUUIDs to UUIDFields (which involves no
transformation in the database) we use a consistent (longer) string.

* Handle legacy SmallUUID links

There are links in the wild for invites and downloads that use
SmallUUIDs. We need to still process those, at least for now.

* Add smalluuid to uuid migrations

* Attempt to fix tests by using different migrations

* Fix actionnetwork migration hex_grouped reference

* Be more explicit about which exceptions we're tracking