Switch to gunicorn for dev purposes (#400)
* Switch to gunicorn for dev purposes

gevent multiprocessing doesn't work with the django dev server. We need
to switch to developing using gunicorn w/gevent directly. I
significantly slimmed-down the gunicorn config from the prod one and
left out a lot of the datadog stuff, but it's still noisy with
datadog attempts. We could probably override `common.apm.Tracer` and
start spitting APM calls to the command line, but that's a future
project.

* Reload on .pdf changes