voter: use filter().first() instead of get (#472)
It's possible for multiple records to share the same
voter id (state, ts, alloy, etc) since those are not
primary keys.  Avoid throwing an exception and use the
first one.

This doesn't address the more general concern of what it
means when there are dups; that will require a more
deliberate cleanup.