Reduce crosstalk between workers (#508)
Celery workers keep trying to send heartbeat responses to each other on
a queue that doesn't exist. We don't need our workers to coordinate with
one another, or to verify the health of each other, so we can disable
that functionality. Will also reduce the amount of unnecessary noise
going across redis.