lob: improve avoidance of duplicate letters (#514)
The old check was racy: if there were already 2 submissions, the get()
would throw an exception.

So, first fix that check to use first() instead.

Second, after we do the (slow) lob submission API call, verify that we
were the first worker to successfully do so.  If not, cancel our submission
and delete our Link record.