register: condition all notifications on lob artifact, not address (#529)
We can have a situation like:

          created_at           |    event_type
-------------------------------+------------------
 2020-09-17 12:06:32.172557+00 | FinishLobConfirm
 2020-09-17 12:06:13.222989+00 | FinishLob
 2020-09-17 12:01:56.532671+00 | Start
 2020-09-17 12:01:21.2716+00   | FinishExternal
 2020-09-17 12:01:07.289831+00 | FinishExternal
 2020-09-17 12:00:52.930191+00 | Start

where the upsell (or other notification) triggers between the second Start
and FinishLob, a window in which request_mailing_address is defined but
there is no lob artifact.  The same might happen if the lob flow is never
completed.

To avoid this, condition all notifications on the final result_item_mail,
and not on the address field.  We don't trigger any notifications until
*after* we populate result_item_mail in generateform.py, with the exception
of the unfinished check, which only sends a text and is not conditional
on the lob vs other completion flow.